+++ himalaya/kernel_ridge/tests/test_sklearn_api_kernel.py
@@ -446,6 +446,23 @@ def test_weighted_kernel_ridge_cv_array_deltas(backend):
         model_1.fit(Ks, Y)
 
 
+@pytest.mark.parametrize('Estimator', [
+    Ridge,
+    RidgeCV,
+    KernelRidge,
+    KernelRidgeCV,
+    MultipleKernelRidgeCV,
+    # WeightedKernelRidge,
+])
+@pytest.mark.parametrize('backend', ALL_BACKENDS)
+def test_n_targets_batch(backend, Estimator):
+    backend = set_backend(backend)
+    Xs, Ks, Y = _create_dataset(backend)
+
+    model = Estimator(solver_params=dict(n_targets_batch=2))
+    model.fit(Xs[0], Y)
+
+
 ###############################################################################
 ###############################################################################
 ###############################################################################